refactor: clean up page.tsx in component stack/page.tsx
This commit represents a significant clean-up of the Home component within page.tsx. The large chunk of the code related to the theme setup and generation of a graphical family tree, based on the go library, was removed.

The removed part was too complex and was affecting the readability of the main component's code. This change improves the maintainability and readability of the component by simplifying it.

If this tree feature still needed in the future, consider extracting it into its own component or service to minimize complexity within the Home component. Meanwhile, attentions could be concentrated on the remaining useEffect which seems to utilize the go library as well.

No references to existing issues were found in relation to this commit.