feat: Removed Baselime template from README.md
We've removed all content related to the Baselime template with OpenTelemetry for Next.js from the README.md. This includes sections on deployment, logging, distributed tracing, usage instructions, and additional resources. This is part of an effort to simplify the project's documentation and focus it more tightly on its core functionality. Further information and resources regarding Baselime and OpenTelemetry will be provided elsewhere in order to not overload the main project documentation.