feat: Update home page links
This update introduces changes to the main "Get Started" and Github links on the homepage. The "Get Started" link now redirects to the homepage ("/") instead of previously doing nothing ("#"). The Github link now accurately refers to the actual Github repository instead of previously doing nothing ("#"). Additionally, the custom SVG for the Github link was replaced with an IconBrandGithub component for a more consistent UI.

No breaking changes or considerations resulting from these changes.